rust: fix building on macOS (issue6801)
The VFS change is copied over from Cargo, and likely to apply to other
platforms as well.

The dirstate change is essentially a replay of 440972d2175d, which was
reverted in e98fd81bb151, part of !383, to silence some clippy
warnings.

+++ rust/hg-core/src/dirstate_tree/on_disk.rs
@@ -415,9 +415,9 @@ impl Node {
 
     fn synthesize_unix_mode(&self) -> u32 {
         let file_type = if self.flags().contains(Flags::MODE_IS_SYMLINK) {
-            libc::S_IFLNK
+            libc::S_IFLNK as u32
         } else {
-            libc::S_IFREG
+            libc::S_IFREG as u32
         };
         let permissions = if self.flags().contains(Flags::MODE_EXEC_PERM) {
             0o755
